rename to SWCloseEvent
